Update redirection links in vercel.json
The redirection links were updated in vercel.json. The purpose of this change was to align all existing paths to a new archive page. This was done to improve the overall site structure, navigation and SEO. With this change all traffic will be redirected to the new English language archive page. Moreover, new redirection entries for volume specific paths and author specific paths were added.